requirements.txt
click==8.1.2
flask==2.1.1
flask_migrate==3.1.0
flask_wtf==1.0.1
flask-restx==0.5.1
flask_login==0.6.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jsonify==0.5
pycryptodome==3.14.1
MarkupSafe==2.1.1
PyJWT==2.3.0
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
python-dotenv==0.20.0
loguru==0.6.0
pandas==1.4.2
email_validator==1.2.0
python-decouple==3.6
simplejson==3.17.6
simple_rest_client==1.1.2
passlib==1.7.4
bcrypt==3.2.0
